+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./commands/effects.js";
+export * from "./commands/help.js";
 export * from "./commands/parse.js";
 export * from "./config/bootstrap.js";
 export * from "./context/repo-graph.js";